# chore(engine): Prefetch inputs of SortMerge operator