--- a/pkg/engine/executor/pipeline_test.go
+++ b/pkg/engine/executor/pipeline_test.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/apache/arrow-go/v18/arrow"
 	"github.com/apache/arrow-go/v18/arrow/csv"
@@ -11,6 +12,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/grafana/loki/v3/pkg/engine/internal/datatype"
+	"github.com/grafana/loki/v3/pkg/util/arrowtest"
 )
 
 // CSVToArrow converts a CSV string to an Arrow record based on the provided schema.
@@ -106,4 +108,113 @@ func TestCSVPipeline(t *testing.T) {
 		_, err = pipeline.Value()
 		require.Equal(t, EOF, err)
 	})
+}
+
+func newInstrumentedPipeline(inner Pipeline) *instrumentedPipeline {
+	return &instrumentedPipeline{
+		inner:     inner,
+		callCount: make(map[string]int),
+	}
+}
+
+type instrumentedPipeline struct {
+	inner     Pipeline
+	callCount map[string]int
+}
+
+// Inputs implements Pipeline.
+func (i *instrumentedPipeline) Inputs() []Pipeline {
+	i.callCount["Inputs"]++
+	return i.inner.Inputs()
+}
+
+// Close implements Pipeline.
+func (i *instrumentedPipeline) Close() {
+	i.callCount["Close"]++
+	i.inner.Close()
+}
+
+// Read implements Pipeline.
+func (i *instrumentedPipeline) Read() error {
+	i.callCount["Read"]++
+	return i.inner.Read()
+}
+
+// Transport implements Pipeline.
+func (i *instrumentedPipeline) Transport() Transport {
+	i.callCount["Transport"]++
+	return i.inner.Transport()
+}
+
+// Value implements Pipeline.
+func (i *instrumentedPipeline) Value() (arrow.Record, error) {
+	i.callCount["Value"]++
+	return i.inner.Value()
+}
+
+var _ Pipeline = (*instrumentedPipeline)(nil)
+
+func Test_prefetchWrapper_Read(t *testing.T) {
+	alloc := memory.NewCheckedAllocator(memory.NewGoAllocator())
+
+	batch1 := arrowtest.Rows{
+		{"message": "log line 1"},
+		{"message": "log line 2"},
+	}
+	batch2 := arrowtest.Rows{
+		{"message": "log line 3"},
+		{"message": "log line 4"},
+	}
+	batch3 := arrowtest.Rows{
+		{"message": "log line 5"},
+	}
+
+	pipeline := NewBufferedPipeline(
+		batch1.Record(alloc, batch1.Schema()),
+		batch2.Record(alloc, batch2.Schema()),
+		batch3.Record(alloc, batch3.Schema()),
+	)
+
+	instrumentedPipeline := newInstrumentedPipeline(pipeline)
+	prefetchingPipeline := newPrefetchingPipeline(t.Context(), instrumentedPipeline)
+
+	require.Equal(t, 0, instrumentedPipeline.callCount["Read"])
+
+	// Read first batch
+	err := prefetchingPipeline.Read()
+	require.NoError(t, err)
+	v, _ := prefetchingPipeline.Value()
+	require.Equal(t, int64(2), v.NumRows())
+
+	time.Sleep(10 * time.Millisecond)                           // ensure that next batch has been prefetched
+	require.Equal(t, 2, instrumentedPipeline.callCount["Read"]) // 1 record consumed + 1 record pre-fetched
+
+	// Read second batch
+	err = prefetchingPipeline.Read()
+	require.NoError(t, err)
+	v, _ = prefetchingPipeline.Value()
+	require.Equal(t, int64(2), v.NumRows())
+
+	time.Sleep(10 * time.Millisecond)                           // ensure that next batch has been prefetched
+	require.Equal(t, 3, instrumentedPipeline.callCount["Read"]) // 2 records consumed + 1 record pre-fetched
+
+	// Read third/last batch
+	err = prefetchingPipeline.Read()
+	require.NoError(t, err)
+	v, _ = prefetchingPipeline.Value()
+	require.Equal(t, int64(1), v.NumRows())
+
+	time.Sleep(10 * time.Millisecond)                           // ensure that next batch has been prefetched
+	require.Equal(t, 4, instrumentedPipeline.callCount["Read"]) // 3 records consumed + 1 EOF pre-fetched
+
+	// Read EOF
+	err = prefetchingPipeline.Read()
+	require.ErrorContains(t, err, EOF.Error())
+
+	time.Sleep(10 * time.Millisecond)                           // ensure that next batch has been prefetched
+	require.Equal(t, 4, instrumentedPipeline.callCount["Read"]) // 3 records + 1 EOF consumed
+
+	// Assert that records are released
+	prefetchingPipeline.Close()
+	alloc.AssertSize(t, 0)
 }