--- a/pkg/engine/executor/sortmerge.go
+++ b/pkg/engine/executor/sortmerge.go
@@ -27,15 +27,17 @@ func NewSortMergePipeline(inputs []Pipeline, order physical.SortOrder, column ph
 		return nil, fmt.Errorf("invalid sort order %v", order)
 	}
 
-	// TODO(chaudum): Pass context from execution engine
-	ctx, cancel := context.WithCancel(context.Background())
+	// TODO(chaudum): Pass context from execution context
+	ctx := context.Background()
+
+	for i := range inputs {
+		inputs[i] = newPrefetchingPipeline(ctx, inputs[i])
+	}
 
 	return &KWayMerge{
 		inputs:     inputs,
 		columnEval: evaluator.newFunc(column),
 		compare:    lessFunc,
-		ctx:        ctx,
-		cancel:     cancel,
 	}, nil
 }
 
@@ -50,11 +52,8 @@ type KWayMerge struct {
 	batches     []arrow.Record
 	exhausted   []bool
 	offsets     []int64
-	fetchers    []chan state
 	columnEval  evalFunc
 	compare     compareFunc[int64]
-	ctx         context.Context
-	cancel      context.CancelFunc
 }
 
 var _ Pipeline = (*KWayMerge)(nil)
@@ -65,14 +64,6 @@ func (p *KWayMerge) Close() {
 	if p.state.batch != nil {
 		p.state.batch.Release()
 	}
-	// Stop pending prefetchers
-	p.cancel()
-	for _, fetcher := range p.fetchers {
-		for range fetcher {
-			// pull remaining items from channel so the goroutine can exit
-		}
-	}
-	// Close child operators
 	for _, input := range p.inputs {
 		input.Close()
 	}
@@ -110,37 +101,9 @@ func (p *KWayMerge) init() {
 	p.batches = make([]arrow.Record, n)
 	p.exhausted = make([]bool, n)
 	p.offsets = make([]int64, n)
-	p.fetchers = make([]chan state, n)
-
-	for i := range n {
-		p.fetchers[i] = make(chan state, 1)
-		go p.prefetch(p.ctx, i) // nolint:errcheck
-	}
 
 	if p.compare == nil {
 		p.compare = func(a, b int64) bool { return a <= b }
-	}
-}
-
-func (p *KWayMerge) prefetch(ctx context.Context, i int) error {
-	// Close channel on exit
-	defer close(p.fetchers[i])
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-			var s state
-			s.err = p.inputs[i].Read()
-			if s.err != nil {
-				p.fetchers[i] <- s
-				return s.err
-			}
-			s.batch, _ = p.inputs[i].Value()
-			s.batch.Retain()
-			p.fetchers[i] <- s
-		}
 	}
 }
 
@@ -158,38 +121,33 @@ start:
 	timestamps := make([]int64, 0, len(p.inputs))
 	inputIndexes := make([]int, 0, len(p.inputs))
 
+loop:
 	for i := range len(p.inputs) {
 		// Skip exhausted inputs
 		if p.exhausted[i] {
-			continue
+			continue loop
 		}
 
 		// Load next batch if it hasn't been loaded yet, or if current one is already fully consumed
 		// Read another batch as long as the input yields zero-length batches.
-		if p.batches[i] == nil || p.offsets[i] == p.batches[i].NumRows() {
+		for p.batches[i] == nil || p.offsets[i] == p.batches[i].NumRows() {
 			// Reset offset
 			p.offsets[i] = 0
 
-			// Release previous batch
-			if p.batches[i] != nil {
-				p.batches[i].Release()
-			}
-
 			// Read from input
-			s := <-p.fetchers[i]
-
-			if s.err != nil {
-				if errors.Is(s.err, EOF) {
+			err := p.inputs[i].Read()
+			if err != nil {
+				if errors.Is(err, EOF) {
 					p.exhausted[i] = true
 					p.batches[i] = nil // remove reference to arrow.Record from slice
-					continue
+					continue loop
 				}
-				return s.err
+				return err
 			}
 
 			// It is safe to use the value from the Value() call, because the error is already checked after the Read() call.
 			// In case the input is exhausted (reached EOF), the return value is `nil`, however, since the flag `p.exhausted[i]` is set, the value will never be read.
-			p.batches[i] = s.batch
+			p.batches[i], _ = p.inputs[i].Value()
 		}
 
 		// Fetch timestamp value at current offset