### Summary

This PR changes how inputs of the `SortMerge` operator are consumed. The inputs are wrapped into the `prefetchWrapper` which consumses the next batch (arrow.Record) in the background while the current batch is processed by the parent node.

### Benchmarks

```
$ gotest -v ./pkg/logql/bench/... -count=10 -test.bench='^BenchmarkLogQL/query=.+\[BACKWARD\]/kind=log/store=dataobj-engine' -test.run=^$ -timeout=30m -benchmem
```

```console
$ benchstat bench_sortbytimestamp_without_prefetch.txt bench_sortbytimestamp_with_prefetch.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/v3/pkg/logql/bench
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                                                                                                                                                              │ bench_sortbytimestamp_without_prefetch.txt │ bench_sortbytimestamp_with_prefetch.txt │
                                                                                                                                                              │                   sec/op                   │     sec/op       vs base                │
LogQL/query={region="ap-southeast-1",_env="dev"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                   770.6m ± 67%     545.5m ± 171%        ~ (p=0.280 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                           1.872 ± 16%      1.409 ±  35%  -24.73% (p=0.004 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                            1.776 ±  3%      1.190 ±   5%  -32.99% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                      1.576 ±  3%      1.043 ±  10%  -33.82% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                               1.477 ±  9%      1.085 ±  16%  -26.51% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                      1.811 ±  3%      1.202 ±  28%  -33.61% (p=0.002 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                       1.869 ±  4%      1.225 ±   4%  -34.47% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                 1.543 ± 11%      1.008 ±  15%  -34.69% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                1.712 ±  7%      1.052 ±  14%  -38.55% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                       3.420 ±  4%      2.497 ±  30%  -26.99% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                        3.971 ± 25%      2.108 ±   5%  -46.91% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                  3.599 ± 41%      2.181 ±  12%  -39.39% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                           1320.2m ± 37%     789.5m ±   8%  -40.19% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                    1.577 ±  9%      1.117 ±  12%  -29.15% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                     1.520 ±  7%      1.045 ±   9%  -31.29% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                             1524.1m ±  4%     904.0m ±  10%  -40.69% (p=0.000 n=10)
geomean                                                                                                                                                                                        1.804            1.187         -34.22%

                                                                                                                                                              │ bench_sortbytimestamp_without_prefetch.txt │  bench_sortbytimestamp_with_prefetch.txt   │
                                                                                                                                                              │             kilobytesProcessed             │ kilobytesProcessed  vs base                │
LogQL/query={region="ap-southeast-1",_env="dev"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                    3.990k ± 0%          5.442k ± 0%  +36.39% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                           3.990k ± 0%          5.442k ± 0%  +36.39% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                            4.005k ± 0%          5.458k ± 0%  +36.28% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                      4.608k ± 0%          5.874k ± 0%  +27.47% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                               5.699k ± 0%          8.839k ± 0%  +55.10% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                      5.699k ± 0%          8.839k ± 0%  +55.10% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                       5.699k ± 0%          8.839k ± 0%  +55.10% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                 5.234k ± 0%          7.905k ± 0%  +51.03% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                5.140k ± 0%          5.571k ± 0%   +8.39% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                       11.68k ± 0%          12.11k ± 0%   +3.68% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                        11.94k ± 0%          12.36k ± 0%   +3.56% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                  12.09k ± 0%          13.08k ± 0%   +8.19% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                             3.368k ± 0%          4.199k ± 0%  +24.67% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                    4.160k ± 0%          5.001k ± 0%  +20.22% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                     4.160k ± 0%          5.001k ± 0%  +20.22% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                               3.864k ± 0%          4.549k ± 0%  +17.73% (p=0.000 n=10)
geomean                                                                                                                                                                                        5.421k               6.910k       +27.48%

                                                                                                                                                              │ bench_sortbytimestamp_without_prefetch.txt │ bench_sortbytimestamp_with_prefetch.txt │
                                                                                                                                                              │               linesProcessed               │ linesProcessed   vs base                │
LogQL/query={region="ap-southeast-1",_env="dev"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                    68.83k ± 0%      114.39k ± 0%  +66.21% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                           68.83k ± 0%      114.39k ± 0%  +66.21% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                            69.34k ± 0%      114.91k ± 0%  +65.72% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1",_env="dev"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                      97.50k ± 0%      143.07k ± 0%  +46.74% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                               68.83k ± 0%      114.39k ± 0%  +66.21% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                      68.83k ± 0%      114.39k ± 0%  +66.21% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                       68.83k ± 0%      114.39k ± 0%  +66.21% (p=0.000 n=10)
LogQL/query={region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                 68.83k ± 0%      114.39k ± 0%  +66.21% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                322.5k ± 0%       372.7k ± 0%  +15.56% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                       1.075M ± 0%       1.125M ± 0%   +4.67% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                        1.104M ± 0%       1.153M ± 0%   +4.45% (p=0.000 n=10)
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                  1.142M ± 0%       1.258M ± 1%  +10.18% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                             68.83k ± 0%      114.39k ± 0%  +66.21% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                    112.3k ± 0%       157.9k ± 0%  +40.56% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                     112.3k ± 0%       157.9k ± 0%  +40.56% (p=0.000 n=10)
LogQL/query={service_name="kubernetes"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                               112.9k ± 0%       158.4k ± 0%  +40.38% (p=0.000 n=10)
geomean                                                                                                                                                                                        143.1k            205.5k       +43.65%

                                                                                                                                                              │ bench_sortbytimestamp_without_prefetch.txt │ bench_sortbytimestamp_with_prefetch.txt  │
                                                                                                                                                              │              postFilterLines               │ postFilterLines  vs base                 │
LogQL/query={region="ap-southeast-1",_env="dev"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                   0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                          0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1",_env="dev"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                           0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1",_env="dev"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                     0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                              0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                     0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                      0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                               0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                      0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                       0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="grafana",_env="prod",_region="us-west-2"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                 0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="kubernetes"}_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                                            0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="kubernetes"}_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                   0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="kubernetes"}_|_detected_level="warn"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                                                    0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
LogQL/query={service_name="kubernetes"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]/kind=log/store=dataobj-engine-8                                                              0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                                                                                                                                                  ²                    +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                                                                                                                              │ bench_sortbytimestamp_without_prefetch.txt │ bench_sortbytimestamp_with_prefetch.txt │
…